examples/live-dashboard/app/worker-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
httpx==0.27.0
fluent-logger==0.10.0
